More lib functions

+++ Programs/Lib/Include/string.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "size_t.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -8,19 +9,19 @@ int strcmp(const char* str1, const char* str2);
 int stricmp(char const *a, char const *b);
 char* strcpy(char *dest, const char *src);
 char* strncpy(char *dest, const char *src, int length);
-unsigned int strlen(const char* str);
+size_t strlen(const char* str);
 char* strcat(char* dest, const char* src);
 char* strncat(char* dest, const char* src, int n);
 
 void stoupper(char* s);
 void stolower(char* s);
 
-void* memcpy(void* dest, const void* src, unsigned int n);
-int memcmp(const void* a, const void* b, unsigned int n);
+void* memcpy(void* dest, const void* src, size_t n);
+int memcmp(const void* a, const void* b, size_t n);
 
 void* memset(void* dest, int val, int n);
-void* memset16(void* dest, unsigned short val, unsigned int n);
-void* memset32(void* dest, unsigned int val, unsigned int n);
+void* memset16(void* dest, unsigned short val, size_t n);
+void* memset32(void* dest, unsigned int val, size_t n);
 
 char* search(const char* str, const char* delim);
 char* strtok_r(char* str, const char* delim, char** saveptr);
